Move the Context key definition and accessor to io.grpc.
